pyproject.toml
[project]
name = "crispr-off-t-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dvc>=3.59.1",
    "dvc-webdav>=3.0.0",
    "hydra-core>=1.3.2",
    "mlflow>=2.22.0",
    "pandas==2.2.3",
    "pre-commit>=4.2.0",
    "pytorch-lightning==2.5.0.post0",
    "seaborn>=0.13.2",
    "transformers==4.47.0",
    "wandb==0.19.1",
]
